Talk proposal - API design introduction
The API design is a topic I find to be interesting and I've learned it
is sometimes hard to both development teams producing and consuming the
APIs.
I will go over ideas and tricks on how to approach designing web APIs,
what to think about and what to watch for, as well as share a lot of my
experience and troubles I've seen as best as it can be done in a short
talk. This is also an introduction into more talks and topics I plan to
cover in future meetups.